Revert "#949 POM up product-api version; update CCM Pgp and CCM Pts request"
This reverts commit 4603fc52017683fc08f2044dde1743f88ba36862.

+++ src/main/java/com/nlmk/kb/server/api/ccm/pts/CcmPtsRequest.java
@@ -69,7 +69,7 @@ public class CcmPtsRequest {
         @NotNull
         @Valid
         private Geometry geometry; // Геометрия
-        private Long orderNum; // Номер заказа
+        private Integer orderNum; // Номер заказа
         private Integer orderPos; // Позиция заказа
         private Integer unionId; // Идентификатор упаковки бунтов
         private List<@Valid Bundle> bundles; // Список бунтов входящих в одну связку
